+++ src/components/LoadScenarioDialog.js
@@ -10,7 +10,6 @@ import Divider from '@material-ui/core/Divider'
 import { makeStyles } from '@material-ui/core/styles'
 import TextField from '@material-ui/core/TextField'
 
-
 const useStyles = makeStyles(theme => ({
   dialog: {
     padding: theme.spacing(2),
@@ -19,32 +18,8 @@ const useStyles = makeStyles(theme => ({
 
 export default (props) => {
   const classes = useStyles()
-  let scenarioRef = React.useRef(null)
 
-  const copyScenario = () => {
-    scenarioRef.current.children[1].children[0].select()
-    document.execCommand('copy')
-    props.toggle()
-    props.toast('Scenario copied to clipboard', 'success')
-  }
-
-  const scenario = {
-    classification: 'UNCLASSIFIED',
-    date: new Date(),
-    data: {
-      buildingLabels: props.data.buildingLabels,
-      combatAirPatrols: props.data.combatAirPatrols,
-      engagementAreas: props.data.engagementAreas,
-      friendlyMarkers: props.data.friendlyMarkers,
-      hostileMarkers: props.data.hostileMarkers,
-      initialPoints: props.data.initialPoints,
-      lines: props.data.lines,
-      polygons: props.data.polygons,
-      restrictedOperatingZones: props.data.restrictedOperatingZones,
-      survivors: props.data.survivors,
-      threatMarkers: props.data.threatMarkers
-    }
-  }
+  const [scenario, setScenario] = React.useState('')
 
   return (
     <Dialog
@@ -54,22 +29,21 @@ export default (props) => {
       onClose={props.toggle}
       maxWidth='lg'
     >
-      <DialogTitle>Save Scenario</DialogTitle>
+      <DialogTitle>Load Scenario</DialogTitle>
       <DialogContent>
-        <DialogContentText>Copy the data below</DialogContentText>
         <TextField
           className={classes.form}
           autoFocus={true}
           fullWidth={true}
           label='Scenario Data'
-          ref={scenarioRef}
-          value={JSON.stringify(scenario)}
-          variant='outlined'
           multiline={true}
+          onChange={event => setScenario(event.target.value)}
+          value={scenario}
+          variant='outlined'          
         />
       </DialogContent>
       <DialogActions>
-        <Button onClick={copyScenario} color='primary'>Copy Scenario</Button>
+        <Button onClick={() => props.submit(scenario)} color='primary'>Load Scenario</Button>
         <Button onClick={props.toggle}>Close</Button>
       </DialogActions>
     </Dialog>